+++ guided_diffusion/scoreVAE_pl_module.py
@@ -242,7 +242,7 @@ class ScoreVAE(pl.LightningModule):
 
     def encode(self, x):
         #compute the parameters of the encoding distribution p_φ(z|x_t)
-        latent_distribution_parameters = self.encoder(x, torch.full(size=(x.size(0),), fill_value=-1).to(self.device)) #self.encoder(x, th.zeros(size=(x.size(0),)).to(self.device))
+        latent_distribution_parameters = self.encoder(x, torch.full(size=(x.size(0),), fill_value=-1).long().to(self.device)) #self.encoder(x, th.zeros(size=(x.size(0),)).to(self.device))
         latent_dim = latent_distribution_parameters.size(1)//2
         mean_z = latent_distribution_parameters[:, :latent_dim]
         log_var_z = latent_distribution_parameters[:, latent_dim:]
